increase header size w/ h4. revert bootstrap col resize
@cjrmbon
`col-md-12` refers to the width of the column, with a page being 12-units wide. So `col-md-12` says "use the whole width of the page".

Setting font size exactly is trickier, but I think we can achieve the desired effect by using a "header" element `<h4>`. To go even bigger try `<h3>`.